[modify] コンテキスト周りのプロジェクト依存関係の調整
