+++ src/components/molecules/Enemy.tsx
@@ -3,7 +3,7 @@
 import { motion } from 'framer-motion'
 import Image from 'next/image'
 import React, { useContext } from 'react'
-import { Context } from '@/app/battle/context'
+import { Context } from '@/game/context'
 import { ActionCharacter } from '@/models/actionCharacter'
 import './enemy.css'
