+++ src/components/molecules/FieldPlayer.tsx
@@ -3,7 +3,7 @@
 import { motion } from 'framer-motion'
 import Image from 'next/image'
 import React, { useContext } from 'react'
-import { BattleState, Context } from '@/app/battle/context'
+import { BattleState, Context } from '@/game/context'
 import { ActionCharacter } from '@/models/actionCharacter'
 
 export type Props = {
